fix: atomic flush incorrect use and clean up code
