+++ common/src/main/java/org/apache/rocketmq/common/config/AbstractRocksDBStorage.java
@@ -17,18 +17,10 @@
 package org.apache.rocketmq.common.config;
 
 import com.google.common.collect.Maps;
-import java.nio.ByteBuffer;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ArrayBlockingQueue;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.Semaphore;
-import java.util.concurrent.ThreadPoolExecutor;
-import java.util.concurrent.TimeUnit;
+import io.netty.buffer.PooledByteBufAllocator;
+import java.nio.charset.StandardCharsets;
 import org.apache.rocketmq.common.ThreadFactoryImpl;
 import org.apache.rocketmq.common.constant.LoggerName;
-import org.apache.rocketmq.common.utils.DataConverter;
 import org.apache.rocketmq.common.utils.ThreadUtils;
 import org.apache.rocketmq.logging.org.slf4j.Logger;
 import org.apache.rocketmq.logging.org.slf4j.LoggerFactory;
@@ -37,7 +29,9 @@ import org.rocksdb.ColumnFamilyHandle;
 import org.rocksdb.ColumnFamilyOptions;
 import org.rocksdb.CompactRangeOptions;
 import org.rocksdb.CompactionOptions;
+import org.rocksdb.CompressionType;
 import org.rocksdb.DBOptions;
+import org.rocksdb.Env;
 import org.rocksdb.FlushOptions;
 import org.rocksdb.LiveFileMetaData;
 import org.rocksdb.Priority;
@@ -49,14 +43,31 @@ import org.rocksdb.Status;
 import org.rocksdb.WriteBatch;
 import org.rocksdb.WriteOptions;
 
-import static org.rocksdb.RocksDB.NOT_FOUND;
+import java.nio.ByteBuffer;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.ArrayBlockingQueue;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.Semaphore;
+import java.util.concurrent.ThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
 
 public abstract class AbstractRocksDBStorage {
     protected static final Logger LOGGER = LoggerFactory.getLogger(LoggerName.ROCKSDB_LOGGER_NAME);
 
+    /**
+     * Direct Jemalloc allocator
+     */
+    public static final PooledByteBufAllocator POOLED_ALLOCATOR = new PooledByteBufAllocator(true);
+
+    public static final byte CTRL_0 = '\u0000';
+    public static final byte CTRL_1 = '\u0001';
+    public static final byte CTRL_2 = '\u0002';
+
     private static final String SPACE = " | ";
 
-    protected String dbPath;
+    protected final String dbPath;
     protected boolean readOnly;
     protected RocksDB db;
     protected DBOptions options;
@@ -71,7 +82,8 @@ public abstract class AbstractRocksDBStorage {
     protected CompactRangeOptions compactRangeOptions;
 
     protected ColumnFamilyHandle defaultCFHandle;
-    protected final List<ColumnFamilyOptions> cfOptions = new ArrayList();
+    protected final List<ColumnFamilyOptions> cfOptions = new ArrayList<>();
+    protected final List<ColumnFamilyHandle> cfHandles = new ArrayList<>();
 
     protected volatile boolean loaded;
     private volatile boolean closed;
@@ -79,13 +91,74 @@ public abstract class AbstractRocksDBStorage {
     private final Semaphore reloadPermit = new Semaphore(1);
     private final ScheduledExecutorService reloadScheduler = ThreadUtils.newScheduledThreadPool(1, new ThreadFactoryImpl("RocksDBStorageReloadService_"));
     private final ThreadPoolExecutor manualCompactionThread = (ThreadPoolExecutor) ThreadUtils.newThreadPoolExecutor(
-            1, 1, 1000 * 60, TimeUnit.MILLISECONDS,
-            new ArrayBlockingQueue(1),
-            new ThreadFactoryImpl("RocksDBManualCompactionService_"),
-            new ThreadPoolExecutor.DiscardOldestPolicy());
+        1, 1, 1000 * 60, TimeUnit.MILLISECONDS,
+        new ArrayBlockingQueue<>(1),
+        new ThreadFactoryImpl("RocksDBManualCompactionService_"),
+        new ThreadPoolExecutor.DiscardOldestPolicy());
 
     static {
         RocksDB.loadLibrary();
+    }
+
+    public AbstractRocksDBStorage(String dbPath) {
+        this.dbPath = dbPath;
+    }
+
+    protected void initOptions() {
+        initWriteOptions();
+        initAbleWalWriteOptions();
+        initReadOptions();
+        initTotalOrderReadOptions();
+        initCompactRangeOptions();
+        initCompactionOptions();
+    }
+
+    /**
+     * Write options for <a href="https://github.com/facebook/rocksdb/wiki/Atomic-flush">Atomic Flush</a>
+     */
+    protected void initWriteOptions() {
+        this.writeOptions = new WriteOptions();
+        this.writeOptions.setSync(false);
+        this.writeOptions.setDisableWAL(true);
+        this.writeOptions.setNoSlowdown(true);
+    }
+
+    protected void initAbleWalWriteOptions() {
+        this.ableWalWriteOptions = new WriteOptions();
+        this.ableWalWriteOptions.setSync(false);
+        this.ableWalWriteOptions.setDisableWAL(false);
+        this.ableWalWriteOptions.setNoSlowdown(true);
+    }
+
+    protected void initReadOptions() {
+        this.readOptions = new ReadOptions();
+        this.readOptions.setPrefixSameAsStart(true);
+        this.readOptions.setTotalOrderSeek(false);
+        this.readOptions.setTailing(false);
+    }
+
+    protected void initTotalOrderReadOptions() {
+        this.totalOrderReadOptions = new ReadOptions();
+        this.totalOrderReadOptions.setPrefixSameAsStart(false);
+        this.totalOrderReadOptions.setTotalOrderSeek(true);
+        this.totalOrderReadOptions.setTailing(false);
+    }
+
+    protected void initCompactRangeOptions() {
+        this.compactRangeOptions = new CompactRangeOptions();
+        this.compactRangeOptions.setBottommostLevelCompaction(CompactRangeOptions.BottommostLevelCompaction.kForce);
+        this.compactRangeOptions.setAllowWriteStall(true);
+        this.compactRangeOptions.setExclusiveManualCompaction(false);
+        this.compactRangeOptions.setChangeLevel(true);
+        this.compactRangeOptions.setTargetLevel(-1);
+        this.compactRangeOptions.setMaxSubcompactions(4);
+    }
+
+    protected void initCompactionOptions() {
+        this.compactionOptions = new CompactionOptions();
+        this.compactionOptions.setCompression(CompressionType.LZ4_COMPRESSION);
+        this.compactionOptions.setMaxSubcompactions(4);
+        this.compactionOptions.setOutputFileSizeLimit(4 * 1024 * 1024 * 1024L);
     }
 
     public boolean hold() {
@@ -101,8 +174,8 @@ public abstract class AbstractRocksDBStorage {
     }
 
     protected void put(ColumnFamilyHandle cfHandle, WriteOptions writeOptions,
-                       final byte[] keyBytes, final int keyLen,
-                       final byte[] valueBytes, final int valueLen) throws RocksDBException {
+        final byte[] keyBytes, final int keyLen,
+        final byte[] valueBytes, final int valueLen) throws RocksDBException {
         if (!hold()) {
             throw new IllegalStateException("rocksDB:" + this + " is not ready");
         }
@@ -118,7 +191,7 @@ public abstract class AbstractRocksDBStorage {
     }
 
     protected void put(ColumnFamilyHandle cfHandle, WriteOptions writeOptions,
-                       final ByteBuffer keyBB, final ByteBuffer valueBB) throws RocksDBException {
+        final ByteBuffer keyBB, final ByteBuffer valueBB) throws RocksDBException {
         if (!hold()) {
             throw new IllegalStateException("rocksDB:" + this + " is not ready");
         }
@@ -159,13 +232,13 @@ public abstract class AbstractRocksDBStorage {
         }
     }
 
-    protected boolean get(ColumnFamilyHandle cfHandle, ReadOptions readOptions,
-                          final ByteBuffer keyBB, final ByteBuffer valueBB) throws RocksDBException {
+    protected int get(ColumnFamilyHandle cfHandle, ReadOptions readOptions, final ByteBuffer keyBB,
+        final ByteBuffer valueBB) throws RocksDBException {
         if (!hold()) {
             throw new IllegalStateException("rocksDB:" + this + " is not ready");
         }
         try {
-            return this.db.get(cfHandle, readOptions, keyBB, valueBB) != NOT_FOUND;
+            return this.db.get(cfHandle, readOptions, keyBB, valueBB);
         } catch (RocksDBException e) {
             LOGGER.error("get Failed. {}, {}", this.dbPath, getStatusError(e));
             throw e;
@@ -175,8 +248,8 @@ public abstract class AbstractRocksDBStorage {
     }
 
     protected List<byte[]> multiGet(final ReadOptions readOptions,
-                                    final List<ColumnFamilyHandle> columnFamilyHandleList,
-                                    final List<byte[]> keys) throws RocksDBException {
+        final List<ColumnFamilyHandle> columnFamilyHandleList,
+        final List<byte[]> keys) throws RocksDBException {
         if (!hold()) {
             throw new IllegalStateException("rocksDB:" + this + " is not ready");
         }
@@ -190,7 +263,8 @@ public abstract class AbstractRocksDBStorage {
         }
     }
 
-    protected void delete(ColumnFamilyHandle cfHandle, WriteOptions writeOptions, byte[] keyBytes) throws RocksDBException {
+    protected void delete(ColumnFamilyHandle cfHandle, WriteOptions writeOptions,
+        byte[] keyBytes) throws RocksDBException {
         if (!hold()) {
             throw new IllegalStateException("rocksDB:" + this + " is not ready");
         }
@@ -204,7 +278,8 @@ public abstract class AbstractRocksDBStorage {
         }
     }
 
-    protected void delete(ColumnFamilyHandle cfHandle, WriteOptions writeOptions, ByteBuffer keyBB) throws RocksDBException {
+    protected void delete(ColumnFamilyHandle cfHandle, WriteOptions writeOptions, ByteBuffer keyBB)
+        throws RocksDBException {
         if (!hold()) {
             throw new IllegalStateException("rocksDB:" + this + " is not ready");
         }
@@ -218,8 +293,8 @@ public abstract class AbstractRocksDBStorage {
         }
     }
 
-    protected void rangeDelete(ColumnFamilyHandle cfHandle, WriteOptions writeOptions,
-                               final byte[] startKey, final byte[] endKey) throws RocksDBException {
+    protected void rangeDelete(ColumnFamilyHandle cfHandle, WriteOptions writeOptions, final byte[] startKey,
+        final byte[] endKey) throws RocksDBException {
         if (!hold()) {
             throw new IllegalStateException("rocksDB:" + this + " is not ready");
         }
@@ -262,16 +337,17 @@ public abstract class AbstractRocksDBStorage {
         });
     }
 
-    protected void open(final List<ColumnFamilyDescriptor> cfDescriptors,
-                        final List<ColumnFamilyHandle> cfHandles) throws RocksDBException {
+    protected void open(final List<ColumnFamilyDescriptor> cfDescriptors) throws RocksDBException {
+        this.cfHandles.clear();
         if (this.readOnly) {
             this.db = RocksDB.openReadOnly(this.options, this.dbPath, cfDescriptors, cfHandles);
         } else {
             this.db = RocksDB.open(this.options, this.dbPath, cfDescriptors, cfHandles);
         }
-        this.db.getEnv().setBackgroundThreads(8, Priority.HIGH);
-        this.db.getEnv().setBackgroundThreads(8, Priority.LOW);
-
+        assert cfDescriptors.size() == cfHandles.size();
+        try (Env env = this.db.getEnv()) {
+            env.setBackgroundThreads(8, Priority.LOW);
+        }
         if (this.db == null) {
             throw new RocksDBException("open rocksdb null");
         }
@@ -293,6 +369,9 @@ public abstract class AbstractRocksDBStorage {
         }
     }
 
+    /**
+     * Close column family handles except the default column family
+     */
     protected abstract void preShutdown();
 
     public synchronized boolean shutdown() {
@@ -310,11 +389,12 @@ public abstract class AbstractRocksDBStorage {
             }
             this.db.cancelAllBackgroundWork(true);
             this.db.pauseBackgroundWork();
-            //The close order is matter.
+            //The close order matters.
             //1. close column family handles
             preShutdown();
 
             this.defaultCFHandle.close();
+
             //2. close column family options.
             for (final ColumnFamilyOptions opt : this.cfOptions) {
                 opt.close();
@@ -332,15 +412,16 @@ public abstract class AbstractRocksDBStorage {
             if (this.totalOrderReadOptions != null) {
                 this.totalOrderReadOptions.close();
             }
-            if (this.options != null) {
-                this.options.close();
-            }
             //4. close db.
             if (db != null && !this.readOnly) {
                 this.db.syncWal();
             }
             if (db != null) {
                 this.db.closeE();
+            }
+            // Close DBOptions after RocksDB instance is closed.
+            if (this.options != null) {
+                this.options.close();
             }
             //5. help gc.
             this.cfOptions.clear();
@@ -360,19 +441,31 @@ public abstract class AbstractRocksDBStorage {
         return true;
     }
 
-    public void flush(final FlushOptions flushOptions) {
+    public void flush(final FlushOptions flushOptions) throws RocksDBException {
+        flush(flushOptions, this.cfHandles);
+    }
+
+    public void flush(final FlushOptions flushOptions, List<ColumnFamilyHandle> columnFamilyHandles) throws RocksDBException {
         if (!this.loaded || this.readOnly || closed) {
             return;
         }
 
         try {
             if (db != null) {
-                this.db.flush(flushOptions);
+                // For atomic-flush, we have to explicitly specify column family handles
+                // See https://github.com/rust-rocksdb/rust-rocksdb/pull/793
+                // and https://github.com/facebook/rocksdb/blob/8ad4c7efc48d301f5e85467105d7019a49984dc8/include/rocksdb/db.h#L1667
+                this.db.flush(flushOptions, columnFamilyHandles);
             }
         } catch (RocksDBException e) {
             scheduleReloadRocksdb(e);
             LOGGER.error("flush Failed. {}, {}", this.dbPath, getStatusError(e));
+            throw e;
         }
+    }
+
+    public void flushWAL() throws RocksDBException {
+        this.db.flushWal(true);
     }
 
     public Statistics getStatistics() {
@@ -441,10 +534,6 @@ public abstract class AbstractRocksDBStorage {
         LOGGER.info("reload rocksdb OK. {}", this.dbPath);
     }
 
-    public void flushWAL() throws RocksDBException {
-        this.db.flushWal(true);
-    }
-
     private String getStatusError(RocksDBException e) {
         if (e == null || e.getStatus() == null) {
             return "null";
@@ -477,13 +566,13 @@ public abstract class AbstractRocksDBStorage {
             Map<Integer, StringBuilder> map = Maps.newHashMap();
             for (LiveFileMetaData metaData : liveFileMetaDataList) {
                 StringBuilder sb = map.computeIfAbsent(metaData.level(), k -> new StringBuilder(256));
-                sb.append(new String(metaData.columnFamilyName(), DataConverter.CHARSET_UTF8)).append(SPACE).
-                        append(metaData.fileName()).append(SPACE).
-                        append("s: ").append(metaData.size()).append(SPACE).
-                        append("a: ").append(metaData.numEntries()).append(SPACE).
-                        append("r: ").append(metaData.numReadsSampled()).append(SPACE).
-                        append("d: ").append(metaData.numDeletions()).append(SPACE).
-                        append(metaData.beingCompacted()).append("\n");
+                sb.append(new String(metaData.columnFamilyName(), StandardCharsets.UTF_8)).append(SPACE).
+                    append(metaData.fileName()).append(SPACE).
+                    append("s: ").append(metaData.size()).append(SPACE).
+                    append("a: ").append(metaData.numEntries()).append(SPACE).
+                    append("r: ").append(metaData.numReadsSampled()).append(SPACE).
+                    append("d: ").append(metaData.numDeletions()).append(SPACE).
+                    append(metaData.beingCompacted()).append("\n");
             }
 
             map.forEach((key, value) -> logger.info("level: {}\n{}", key, value.toString()));
@@ -492,11 +581,9 @@ public abstract class AbstractRocksDBStorage {
             String indexesAndFilterBlockMemUsage = this.db.getProperty("rocksdb.estimate-table-readers-mem");
             String memTableMemUsage = this.db.getProperty("rocksdb.cur-size-all-mem-tables");
             String blocksPinnedByIteratorMemUsage = this.db.getProperty("rocksdb.block-cache-pinned-usage");
-            logger.info("MemUsage. blockCache: {}, indexesAndFilterBlock: {}, memtable: {}, blocksPinnedByIterator: {}",
-                    blockCacheMemUsage, indexesAndFilterBlockMemUsage, memTableMemUsage, blocksPinnedByIteratorMemUsage);
-        } catch (Exception e) {
-            logger.error("statRocksdb Failed. {}", this.dbPath, e);
-            throw new RuntimeException(e);
+            logger.info("MemUsage. blockCache: {}, indexesAndFilterBlock: {}, MemTable: {}, blocksPinnedByIterator: {}",
+                blockCacheMemUsage, indexesAndFilterBlockMemUsage, memTableMemUsage, blocksPinnedByIteratorMemUsage);
+        } catch (Exception ignored) {
         }
     }
 }
