+++ store/src/main/java/org/apache/rocketmq/store/rocksdb/RocksDBOptionsFactory.java
@@ -16,7 +16,7 @@
  */
 package org.apache.rocketmq.store.rocksdb;
 
-import org.apache.rocketmq.common.config.ConfigRocksDBStorage;
+import org.apache.rocketmq.common.config.ConfigHelper;
 import org.apache.rocketmq.store.MessageStore;
 import org.rocksdb.BlockBasedTableConfig;
 import org.rocksdb.BloomFilter;
@@ -71,7 +71,7 @@ public class RocksDBOptionsFactory {
                 setTableFormatConfig(blockBasedTableConfig).
                 setMemTableConfig(new SkipListMemTableConfig()).
                 setCompressionType(CompressionType.LZ4_COMPRESSION).
-                setBottommostCompressionType(CompressionType.ZSTD_COMPRESSION).
+                setBottommostCompressionType(CompressionType.LZ4_COMPRESSION).
                 setNumLevels(7).
                 setCompactionStyle(CompactionStyle.UNIVERSAL).
                 setCompactionOptionsUniversal(compactionOption).
@@ -134,7 +134,7 @@ public class RocksDBOptionsFactory {
         Statistics statistics = new Statistics();
         statistics.setStatsLevel(StatsLevel.EXCEPT_DETAILED_TIMERS);
         return options.
-                setDbLogDir(ConfigRocksDBStorage.getDBLogDir()).
+                setDbLogDir(ConfigHelper.getDBLogDir()).
                 setInfoLogLevel(InfoLogLevel.INFO_LEVEL).
                 setWalRecoveryMode(WALRecoveryMode.PointInTimeRecovery).
                 setManualWalFlush(true).
@@ -144,9 +144,9 @@ public class RocksDBOptionsFactory {
                 setCreateIfMissing(true).
                 setCreateMissingColumnFamilies(true).
                 setMaxOpenFiles(-1).
-                setMaxLogFileSize(1 * SizeUnit.GB).
+                setMaxLogFileSize(SizeUnit.GB).
                 setKeepLogFileNum(5).
-                setMaxManifestFileSize(1 * SizeUnit.GB).
+                setMaxManifestFileSize(SizeUnit.GB).
                 setAllowConcurrentMemtableWrite(false).
                 setStatistics(statistics).
                 setAtomicFlush(true).
